requirements.txt
google-genai==1.66.0
opencv-python==4.13.0.92
python-dotenv==1.2.2
Pillow==12.1.1
tqdm==4.67.3